`fp-ethereum` refactor
